+++ Demo/iOS/Pulse_Demo_iOSApp.swift
@@ -9,10 +9,10 @@ import PulseUI
 @main
 struct Pulse_Demo_iOSApp: App {
     var body: some Scene {
-//        let _ = testProxy()
+        let _ = testProxy()
         WindowGroup {
             NavigationView {
-                ConsoleView(store: .demo)
+                ConsoleView(store: .shared)
             }
         }
     }