+++ Tests/PulseUITests/ShareStoreViewModelTests.swift
@@ -1,0 +1,15 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2020–2023 Alexander Grebenyuk (github.com/kean).
+
+import XCTest
+@testable import Pulse
+@testable import PulseUI
+
+final class ShareStoreViewModelTests: XCTestCase {
+    var sut: ShareStoreViewModel
+
+    func testSharingMessagesFromCurrentSession() throws {
+
+    }
+}